requirements.txt
bedrock-agentcore
strands-agents
strands-agents[otel]
aws-opentelemetry-distro
boto3
pandas